test: decouple the shared test scaffolding from the uvpacket example
uvpacket is an **experimental example built on this crate, not a
feature of it**. `tests/common/channel.rs` nonetheless imported
`mfsk_core::uvpacket::Mode` at file scope, which made **every one of
the 43 test binaries that write `mod common;`** require the
`uvpacket` feature to compile — the FT8, FT4 and FST4 tests included,
which have nothing to do with it. Verified on an untouched file:
before this change, `cargo test -p mfsk-core --no-default-features
--features std,ft8,fft-rustfft --test ft8_streaming_decode` failed
with `unresolved import mfsk_core::uvpacket`.

It was not only the `Mode` parameter. `K_INFO = 101`,
`SYMBOL_RATE_HZ = 1200` and `BITS_PER_SYMBOL = 2` are uvpacket's own
π/4-DQPSK PHY numbers, and all five callers of
`awgn_sigma_for_eb_n0_info` are uvpacket tests — the helper was never
shared in any meaningful sense, only shared-by-location.

Moved wholesale to `common/uvpacket_channel.rs`, gated
`#[cfg(feature = "uvpacket")]`. What stays in `common/channel.rs` is
genuinely protocol-agnostic: `AwgnChannel`, `RayleighFlatChannel`,
`signal_power`, `SAMPLE_RATE_HZ`.

**Also fixes a defect this PR's predecessor (#272) introduced.** That
commit put `#![cfg(any(feature = "fft-rustfft", feature =
"fft-extern"))]` on `common_selftest.rs`, and under
`cargo test -p mfsk-core --no-default-features --test
common_selftest` it silently reduced the binary to "0 passed; ok" —
the exact green-lie the restructure exists to eliminate,
reintroduced by the commit meant to make the pass count honest. The
gate is gone; an unsupported feature set now fails to *compile*,
which is loud. The gating that remains is per-module and truthful:

    without uvpacket   20 self-tests
    with uvpacket      23 self-tests

Full suite under `MFSK_REQUIRE_CORPUS=1` green (79 binaries);
`scripts/pre-push-check.sh` clean.
