+++ mfsk-core/tests/uvpacket_fm_channel.rs
@@ -26,7 +26,8 @@ use mfsk_core::uvpacket::{AUDIO_CENTRE_HZ, Mode, rx, tx};
 #[allow(dead_code)]
 mod common;
 use common::air_channel::{FmChannel, PhaseFadingModel};
-use common::channel::{awgn_sigma_for_eb_n0_info, signal_power};
+use common::channel::signal_power;
+use common::uvpacket_channel::awgn_sigma_for_eb_n0_info;
 
 const N_BLOCKS: u8 = 4;
