+++ mfsk-core/tests/uvpacket_multichannel.rs
@@ -17,7 +17,8 @@ use mfsk_core::uvpacket::{Mode, tx};
 // `tests/common_selftest.rs`.
 #[allow(dead_code)]
 mod common;
-use common::channel::{AwgnChannel, awgn_sigma_for_eb_n0_info, signal_power};
+use common::channel::{AwgnChannel, signal_power};
+use common::uvpacket_channel::awgn_sigma_for_eb_n0_info;
 
 fn default_fec_opts() -> FecOpts<'static> {
     FecOpts {
