+++ mfsk-core/tests/uvpacket_per_modes_sweep.rs
@@ -28,7 +28,8 @@ use mfsk_core::uvpacket::{AUDIO_CENTRE_HZ, Mode, rx, tx};
 #[allow(dead_code)]
 mod common;
 use common::air_channel::{FmChannel, PhaseFadingModel, SsbChannel};
-use common::channel::{AwgnChannel, RayleighFlatChannel, awgn_sigma_for_eb_n0_info, signal_power};
+use common::channel::{AwgnChannel, RayleighFlatChannel, signal_power};
+use common::uvpacket_channel::awgn_sigma_for_eb_n0_info;
 
 const N_BLOCKS: u8 = 4;
 const N_TRIALS: usize = 30;
